ptychography/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ptyupdate"
description = "Ptychography pr/ob update example"
version = "0.0.0"
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "scipy",
  "h5py",
  "mpi4py",
  "matplotlib",
  "pyzmq",
  "pillow",
  "pyfftw",
  "cupy-cuda12x",
  "pytest"
]

[tool.setuptools.package-data]
"ptyupdate" = ["*.cu", "*.cuh"]